Lab Project 3(Duplicated Code example 1) - Description: The above code represents an activity in a visual novel application. This activity displays game scenarios and provides choices to the user. The code has been improved by removing code duplicates and simplifying the logic structure.
- Methods:

getPageData(): This method returns a PageData object that contains image, question, and button text information based on the received page number as a parameter.
getNextPage(): This method returns the next page number based on the current page and which button the user clicked.

- Conclusion:
The code has been improved by removing code duplicates and simplifying the scenario selection logic. These improvements enhance code cleanliness and readability, making it easier to maintain and develop the code in the future.